requirements.txt
braceexpand==0.1.7
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pillow==10.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.1.0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scikit-metrics==0.1.0
scipy==1.14.1
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.11
torch==2.5.1+rocm6.2
torchaudio==2.5.1+rocm6.2
torchvision==0.20.1+rocm6.2
tqdm==4.66.6
typing_extensions==4.9.0
urllib3==2.2.3
webdataset==0.2.100
